+++ packages/forma-36-website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.75](https://github.com/contentful/forma-36/compare/forma-36-website@0.9.74...forma-36-website@0.9.75) (2020-11-24)
+
+**Note:** Version bump only for package forma-36-website
+
+
+
+
+
 ## [0.9.74](https://github.com/contentful/forma-36/compare/forma-36-website@0.9.73...forma-36-website@0.9.74) (2020-11-23)
 
 **Note:** Version bump only for package forma-36-website